Move Streamly.String to Streamly.Data.String.
