+++ examples/FileSinkServer.hs
@@ -8,7 +8,7 @@
 import System.Environment (getArgs)
 
 import Streamly
-import Streamly.String
+import Streamly.Data.String
 import qualified Streamly.FileSystem.Handle as FH
 import qualified Streamly.Memory.Array as A
 import qualified Streamly.Network.Socket as NS
